Cache and Stale time:
Images are fetched and potentially cached by the browser. Neither react nor tanstack-query are involved.
staleTime is the time after which cached data is stale and background fetching is allowed - not triggered - in case of refocus or remount. Default is 0 ms
gcTime or cahceTime in v4 is the time after which cached data is garbage collected and hard fetching is launched. Default is 5 * 60 * 1000 ms